+++ game-app/src/app/components/icons/DribbleIcon.tsx
@@ -1,0 +1,19 @@
+const DribbleIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 40 40"
+  >
+    <path
+      d="M17.8462 15.4858L9.41709 7.05664L4 12.4737L11.5263 20L4 27.5263L9.41709 32.9434L17.8462 24.5142C20.3393 22.0211 20.3393 17.9789 17.8462 15.4858Z"
+      fill="currentColor"
+    />
+    <path
+      d="M34.1301 15.4858L25.701 7.05664L20.2839 12.4737L27.8102 20L20.2839 27.5263L25.701 32.9434L34.1301 24.5142C36.6233 22.0211 36.6233 17.9789 34.1301 15.4858Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+export default DribbleIcon;
