+++ game-app/src/app/components/icons/index.ts
@@ -1,0 +1,4 @@
+import DribbleIcon from "./DribbleIcon";
+import StaminaIcon from "./StaminaIcon";
+
+export { DribbleIcon, StaminaIcon };
