+++ game-app/src/stories/CardAttribute.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CardAttribute from "../app/components/CardAttribute";
+import CardAttribute from "../app/components/card/CardAttribute";
 
 const meta = {
   title: "CardAttribute",
